Tile / Tilemap changes
A set of changes to define a stricter contract between
tilemap producers/consumers. Existing compatibility has been retained as much as possible.
This set of changes allows a FUTURE TileLayer implementation to be much more liberal on how it stores information. It should be reasonable to expect (minus special per-tile callbacks) to be able to store a cell/tile in 16-24 bytes of data as opposed to the current 130+ (in V8) for separate objects.
These changes also refine documentation and expose some bugs such as `A*` use the `properties` of a Tile object (these properties are per-tile-index/map and not per-tile/layer); this also led to stale information if an index was updated.
Internal structure documentation has been increased and some properties (eg. data) have been explicitly declared as protected.
The getTileRef-style functions allow tile objects to be cached/re-used when rehydrated from a deeper store.
The use of an internal flags field is for future binary map encoding; individual property access is slightly increased but the overall checks (ie. when checking of if a tile collides) are actually decreased (this is not a performance argument either way).

Parser
- Does not generate Phaser.Tile itself
- CSV now unified with Layer creation

Tilemap
- Fixed shuffle (back to original)
- Updated documentation

Tile
- Added `properties` behind a read-only accessor
- Updated documentation to reflect `properties` vs `indexProperties`.
- Added `Tile#debug` documentation

p2/World

- Updated to use new TileLayer interface

TileLayer
- Per-side collides now honored when calculating faces
- Corrected @property documentation to @member
- renderDebug can now render all tiles that `collides`, even if they do now have a set face

TilemapLayer
- Updated getRayCastTiles to new TileLayer

World
- Updated to new tile intersection

Parser
- Documented JSON/CSV specific quirks

TilemapLayer

- Reverted to correctly display "0" index tiles
  - They are simply "not added" by the JSON format

TileLayer
- Minor documentation updates / variable fixes

Tile
- Fixed bug with "bottom"

TilemapLayer

- Added edge-render mode

Tile

- Added "disconnected" property support

Tile

- Updated Physics to work with changes