Point - angleSq fixes and documentation updates
The math for angleSq has been "flattened" out. I _believe_ the math for
angleSq is correct (assuming both the original code was fundamentally
correct and I didn't botch the conversion), but it needs to be rechecked.

Potential breaking change: The formula in Phaser.Point.angle is "reversed"
so that it follows `Phaser.Point.angle(from, to)` like `from.angle(to)`
(where from is a Point instance); I believe that the forumla is now
"right" wrt the documentation and the original/expected behavior, but it
needs to be verified.

Expanded the documentation wrt what "angle" means in context.

Add the definition of 'Point-like' to the class documentation and altered
some documentation around such.